new run script

+++ run.sh
@@ -6,17 +6,20 @@ echo "=> Using weave version: $VERSION"
 echo "=> Using docker binary:"
 docker version
 
-WEAVE_IMAGES=$(docker images | grep weaveworks/weave | wc -l)
-if [ ${WEAVE_IMAGES} -eq "0" ]; then
-    echo "=> Setting up weave images"
-    /weave --local setup
-fi
-
 if [ "${WEAVE_LAUNCH}" = "**None**" ]; then
     echo "WEAVE_LAUNCH is **None**. Not running 'weave launch'"
 else
-    echo "=> Resetting weave on the node"
-    /weave --local reset
+    if [ ! -f "/.weave_launched" ]; then
+        WEAVE_IMAGES=$(docker images | grep weaveworks/weave | wc -l)
+        if [ ${WEAVE_IMAGES} -eq "0" ]; then
+            echo "=> Setting up weave images"
+            /weave --local setup
+        fi
+
+        echo "=> Resetting weave on the node"
+        /weave --local reset
+    fi
+
     if [ ! -z "${WEAVE_PASSWORD}" ]; then
         echo "=> Running: weave launch -password XXXXXX ${WEAVE_LAUNCH}"
         /weave --local launch -password ${WEAVE_PASSWORD} ${WEAVE_LAUNCH} || true
@@ -30,6 +33,8 @@ fi
 
 echo "=> Current weave router status"
 /weave --local status
+
+touch /.weave_launched
 
 docker logs -f weave &
